+++ cli/src/__tests__/shared-common-env-inject.test.ts
@@ -227,109 +227,7 @@ inject_env_vars_local mock_upload mock_run "MSG=it'\\''s a test"
   });
 });
 
-// ── _extract_json_field ────────────────────────────────────────────────
-
-describe("_extract_json_field", () => {
-  it("should extract a simple string field", () => {
-    const result = runBash(`
-_extract_json_field '{"name":"test"}' "d['name']"
-`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("test");
-  });
-
-  it("should extract a nested field", () => {
-    const result = runBash(`
-_extract_json_field '{"server":{"ip":"10.0.0.1"}}' "d['server']['ip']"
-`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("10.0.0.1");
-  });
-
-  it("should extract a numeric field", () => {
-    const result = runBash(`
-_extract_json_field '{"status_code":200}' "d['status_code']"
-`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("200");
-  });
-
-  it("should return default value when field is missing", () => {
-    const result = runBash(`
-_extract_json_field '{"name":"test"}' "d['missing']" "fallback"
-`);
-    expect(result.stdout).toBe("fallback");
-  });
-
-  it("should return default value for invalid JSON", () => {
-    const result = runBash(`
-_extract_json_field 'not json' "d['key']" "default_val"
-`);
-    expect(result.stdout).toBe("default_val");
-  });
-
-  it("should return empty string when no default and field missing", () => {
-    const result = runBash(`
-_extract_json_field '{"a":1}' "d['b']"
-`);
-    // Default is empty string
-    expect(result.stdout).toBe("");
-  });
-
-  it("should handle array indexing", () => {
-    const result = runBash(`
-_extract_json_field '{"items":["first","second"]}' "d['items'][0]"
-`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("first");
-  });
-
-  it("should handle boolean values", () => {
-    const result = runBash(`
-_extract_json_field '{"active":true}' "d['active']"
-`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("true");
-  });
-
-  it("should handle null values with default", () => {
-    const result = runBash(`
-_extract_json_field '{"ip":null}' "d['ip']" "none"
-`);
-    expect(result.stdout).toBe("none");
-  });
-
-  it("should extract from deeply nested response structures", () => {
-    const result = runBash(`
-_extract_json_field '{"instance":{"network":{"v4":[{"ip_address":"192.168.1.100"}]}}}' "d['instance']['network']['v4'][0]['ip_address']"
-`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("192.168.1.100");
-  });
-
-  it("should handle empty JSON object with default", () => {
-    const result = runBash(`
-_extract_json_field '{}' "d['status']" "unknown"
-`);
-    expect(result.stdout).toBe("unknown");
-  });
-
-  it("should handle values with spaces", () => {
-    const result = runBash(`
-_extract_json_field '{"desc":"hello world"}' "d['desc']"
-`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("hello world");
-  });
-
-  it("should handle empty string value", () => {
-    const result = runBash(`
-_extract_json_field '{"ip":""}' "d['ip']"
-`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("");
-  });
-});
+// _extract_json_field tests are in shared-common-json-extraction.test.ts
 
 // ── check_ssh_key_by_fingerprint ───────────────────────────────────────
 
@@ -398,183 +296,7 @@ check_ssh_key_by_fingerprint mock_api "/ssh_keys" "test" 2>&1 | head -1
   });
 });
 
-// ── opencode_install_cmd ───────────────────────────────────────────────
+// opencode_install_cmd tests are in shared-common-logging-utils.test.ts
 
-describe("opencode_install_cmd", () => {
-  it("should output a non-empty string", () => {
-    const result = runBash(`opencode_install_cmd`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout.length).toBeGreaterThan(0);
-  });
-
-  it("should contain architecture detection", () => {
-    const result = runBash(`opencode_install_cmd`);
-    expect(result.stdout).toContain("uname -m");
-    expect(result.stdout).toContain("aarch64");
-    expect(result.stdout).toContain("arm64");
-  });
-
-  it("should contain OS detection", () => {
-    const result = runBash(`opencode_install_cmd`);
-    expect(result.stdout).toContain("uname -s");
-    expect(result.stdout).toContain("tr A-Z a-z");
-  });
-
-  it("should download to a temp file instead of piping curl|tar", () => {
-    const result = runBash(`opencode_install_cmd`);
-    // The robust install pattern downloads to a file first
-    expect(result.stdout).toContain("/tmp/opencode-install");
-    expect(result.stdout).toContain("curl -fsSL -o");
-  });
-
-  it("should download from github releases", () => {
-    const result = runBash(`opencode_install_cmd`);
-    expect(result.stdout).toContain("github.com/anomalyco/opencode/releases");
-  });
-
-  it("should add to PATH in both .bashrc and .zshrc", () => {
-    const result = runBash(`opencode_install_cmd`);
-    expect(result.stdout).toContain(".bashrc");
-    expect(result.stdout).toContain(".zshrc");
-    expect(result.stdout).toContain(".opencode/bin");
-  });
-
-  it("should create install directory with mkdir -p", () => {
-    const result = runBash(`opencode_install_cmd`);
-    expect(result.stdout).toContain("mkdir -p");
-    expect(result.stdout).toContain("$HOME/.opencode/bin");
-  });
-
-  it("should clean up temp files after install", () => {
-    const result = runBash(`opencode_install_cmd`);
-    expect(result.stdout).toContain("rm -rf /tmp/opencode-install");
-  });
-
-  it("should use tar to extract the downloaded archive", () => {
-    const result = runBash(`opencode_install_cmd`);
-    expect(result.stdout).toContain("tar xzf");
-  });
-
-  it("should export PATH at the end for immediate use", () => {
-    const result = runBash(`opencode_install_cmd`);
-    expect(result.stdout).toContain('export PATH="$HOME/.opencode/bin:$PATH"');
-  });
-
-  it("should not use curl progress bar (no -#)", () => {
-    const result = runBash(`opencode_install_cmd`);
-    // The command uses -fsSL (silent) not -# (progress bar)
-    expect(result.stdout).toContain("curl -fsSL");
-    // Ensure the actual download uses -o flag not piping
-    const downloadPart = result.stdout.split("curl -fsSL -o")[1] || "";
-    expect(downloadPart.length).toBeGreaterThan(0);
-  });
-});
-
-// ── track_temp_file + cleanup_temp_files ───────────────────────────────
-
-describe("track_temp_file and cleanup_temp_files", () => {
-  it("should track and clean up a temporary file", () => {
-    const dir = createTempDir();
-    const tempFile = join(dir, "secret.txt");
-    writeFileSync(tempFile, "sensitive data");
-
-    try {
-      const result = runBash(`
-track_temp_file "${tempFile}"
-cleanup_temp_files
-if [ -f "${tempFile}" ]; then echo "EXISTS"; else echo "CLEANED"; fi
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("CLEANED");
-    } finally {
-      rmSync(dir, { recursive: true, force: true });
-    }
-  });
-
-  it("should track multiple temp files", () => {
-    const dir = createTempDir();
-    const file1 = join(dir, "a.txt");
-    const file2 = join(dir, "b.txt");
-    writeFileSync(file1, "data1");
-    writeFileSync(file2, "data2");
-
-    try {
-      const result = runBash(`
-track_temp_file "${file1}"
-track_temp_file "${file2}"
-cleanup_temp_files
-F1="exists"; F2="exists"
-[ -f "${file1}" ] || F1="cleaned"
-[ -f "${file2}" ] || F2="cleaned"
-echo "\$F1 \$F2"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("cleaned cleaned");
-    } finally {
-      rmSync(dir, { recursive: true, force: true });
-    }
-  });
-
-  it("should handle cleanup of non-existent files gracefully", () => {
-    const result = runBash(`
-track_temp_file "/tmp/nonexistent-spawn-test-file-$$"
-cleanup_temp_files
-echo "OK"
-`);
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("OK");
-  });
-
-  it("should preserve exit code through cleanup", () => {
-    const result = runBash(`
-(
-  exit 42
-)
-cleanup_temp_files
-echo "EXIT=\$?"
-`);
-    // cleanup_temp_files preserves the exit code
-    expect(result.stdout).toContain("EXIT=42");
-  });
-});
-
-// ── validate_resource_name ─────────────────────────────────────────────
-
-describe("validate_resource_name", () => {
-  it("should accept valid resource names", () => {
-    const result = runBash(`validate_resource_name "my-resource-123"`);
-    expect(result.exitCode).toBe(0);
-  });
-
-  it("should accept names with underscores", () => {
-    const result = runBash(`validate_resource_name "my_resource"`);
-    expect(result.exitCode).toBe(0);
-  });
-
-  it("should reject empty string", () => {
-    const result = runBash(`validate_resource_name ""`);
-    expect(result.exitCode).not.toBe(0);
-  });
-
-  it("should reject names with spaces", () => {
-    const result = runBash(`validate_resource_name "my resource"`);
-    expect(result.exitCode).not.toBe(0);
-  });
-
-  it("should reject names with shell metacharacters", () => {
-    const result = runBash(`validate_resource_name "test;rm -rf /"`);
-    expect(result.exitCode).not.toBe(0);
-  });
-
-  it("should reject names exceeding max length", () => {
-    const longName = "a".repeat(64);
-    const result = runBash(`validate_resource_name "${longName}"`);
-    expect(result.exitCode).not.toBe(0);
-  });
-
-  it("should accept names at exactly max length (63)", () => {
-    const name = "a".repeat(63);
-    const result = runBash(`validate_resource_name "${name}"`);
-    expect(result.exitCode).toBe(0);
-  });
-});
+// track_temp_file/cleanup_temp_files tests are in shared-common-logging-utils.test.ts
+// validate_resource_name tests are in shared-common-validators.test.ts
