+++ cli/src/__tests__/shared-common-error-polling.test.ts
@@ -293,323 +293,5 @@ describe("extract_api_error_message", () => {
   });
 });
 
-// ── _extract_json_field (additional edge cases) ────────────────────────
-
-describe("_extract_json_field edge cases", () => {
-  it("should return default when python expression raises KeyError", () => {
-    const result = runBash(
-      `_extract_json_field '{"a":1}' "d['nonexistent']" "default_val"`
-    );
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("default_val");
-  });
-
-  it("should return default when python expression raises TypeError on None", () => {
-    const result = runBash(
-      `_extract_json_field '{"a":null}' "d['a']['nested']" "fallback"`
-    );
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("fallback");
-  });
-
-  it("should handle zero value correctly (not treated as default)", () => {
-    const result = runBash(
-      `_extract_json_field '{"count":0}' "d['count']" "default"`
-    );
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("0");
-  });
-
-  it("should handle False value correctly", () => {
-    const result = runBash(
-      `_extract_json_field '{"enabled":false}' "d['enabled']" "default"`
-    );
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("false");
-  });
-});
-
-// ── generic_wait_for_instance ──────────────────────────────────────────
-
-describe("generic_wait_for_instance", () => {
-  describe("successful polling", () => {
-    it("should detect target status and extract IP on first attempt", () => {
-      const result = runBash(`
-# Mock API function that returns active status with IP
-mock_api() {
-  echo '{"instance":{"status":"active","ip":"10.0.0.1"}}'
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/instances/123" "active" \\
-  "d['instance']['status']" "d['instance']['ip']" \\
-  TEST_IP "Instance" 3
-
-echo "IP=$TEST_IP"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("IP=10.0.0.1");
-    });
-
-    it("should poll until target status is reached", () => {
-      const result = runBash(`
-# Counter file tracks how many times API is called
-COUNTER_FILE=$(mktemp)
-echo "0" > "$COUNTER_FILE"
-
-mock_api() {
-  local count
-  count=$(cat "$COUNTER_FILE")
-  count=$((count + 1))
-  echo "$count" > "$COUNTER_FILE"
-
-  if [[ "$count" -lt 3 ]]; then
-    echo '{"server":{"status":"provisioning","ip":""}}'
-  else
-    echo '{"server":{"status":"running","ip":"192.168.1.50"}}'
-  fi
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/servers/abc" "running" \\
-  "d['server']['status']" "d['server']['ip']" \\
-  SERVER_IP "Server" 5
-
-echo "IP=$SERVER_IP"
-rm -f "$COUNTER_FILE"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("IP=192.168.1.50");
-    });
-
-    it("should export the IP to the specified variable name", () => {
-      const result = runBash(`
-mock_api() {
-  echo '{"vm":{"state":"active","address":"203.0.113.5"}}'
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/vms/1" "active" \\
-  "d['vm']['state']" "d['vm']['address']" \\
-  MY_CUSTOM_VAR "VM" 2
-
-echo "RESULT=$MY_CUSTOM_VAR"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("RESULT=203.0.113.5");
-    });
-  });
-
-  describe("timeout behavior", () => {
-    it("should fail after max_attempts when status never reaches target", () => {
-      const result = runBash(`
-mock_api() {
-  echo '{"instance":{"status":"provisioning","ip":""}}'
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/instances/123" "active" \\
-  "d['instance']['status']" "d['instance']['ip']" \\
-  TEST_IP "Instance" 3
-`);
-      expect(result.exitCode).not.toBe(0);
-      expect(result.stderr).toContain("did not become active");
-    });
-
-    it("should use default max_attempts of 60 when not specified", () => {
-      // We can't test 60 iterations, but verify the parameter is accepted
-      const result = runBash(`
-mock_api() {
-  echo '{"instance":{"status":"active","ip":"10.0.0.1"}}'
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/instances/123" "active" \\
-  "d['instance']['status']" "d['instance']['ip']" \\
-  TEST_IP "Instance"
-
-echo "IP=$TEST_IP"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("IP=10.0.0.1");
-    });
-  });
-
-  describe("IP extraction edge cases", () => {
-    it("should keep polling when status matches but IP is empty", () => {
-      const result = runBash(`
-COUNTER_FILE=$(mktemp)
-echo "0" > "$COUNTER_FILE"
-
-mock_api() {
-  local count
-  count=$(cat "$COUNTER_FILE")
-  count=$((count + 1))
-  echo "$count" > "$COUNTER_FILE"
-
-  if [[ "$count" -lt 3 ]]; then
-    # Status is active but IP is empty (still allocating)
-    echo '{"instance":{"status":"active","ip":""}}'
-  else
-    echo '{"instance":{"status":"active","ip":"10.0.0.5"}}'
-  fi
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/instances/1" "active" \\
-  "d['instance']['status']" "d['instance']['ip']" \\
-  TEST_IP "Instance" 5
-
-echo "IP=$TEST_IP"
-rm -f "$COUNTER_FILE"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("IP=10.0.0.5");
-    });
-
-    it("should handle deeply nested IP field", () => {
-      const result = runBash(`
-mock_api() {
-  echo '{"droplet":{"status":"active","networks":{"v4":[{"ip_address":"198.51.100.1"}]}}}'
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/droplets/42" "active" \\
-  "d['droplet']['status']" "d['droplet']['networks']['v4'][0]['ip_address']" \\
-  DROPLET_IP "Droplet" 2
-
-echo "IP=$DROPLET_IP"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("IP=198.51.100.1");
-    });
-  });
-
-  describe("API error handling", () => {
-    it("should continue polling when API returns error", () => {
-      const result = runBash(`
-COUNTER_FILE=$(mktemp)
-echo "0" > "$COUNTER_FILE"
-
-mock_api() {
-  local count
-  count=$(cat "$COUNTER_FILE")
-  count=$((count + 1))
-  echo "$count" > "$COUNTER_FILE"
-
-  if [[ "$count" -lt 2 ]]; then
-    # API fails
-    return 1
-  fi
-  echo '{"instance":{"status":"active","ip":"10.0.0.3"}}'
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/instances/1" "active" \\
-  "d['instance']['status']" "d['instance']['ip']" \\
-  TEST_IP "Instance" 5
-
-echo "IP=$TEST_IP"
-rm -f "$COUNTER_FILE"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("IP=10.0.0.3");
-    });
-
-    it("should continue polling when API returns invalid JSON", () => {
-      const result = runBash(`
-COUNTER_FILE=$(mktemp)
-echo "0" > "$COUNTER_FILE"
-
-mock_api() {
-  local count
-  count=$(cat "$COUNTER_FILE")
-  count=$((count + 1))
-  echo "$count" > "$COUNTER_FILE"
-
-  if [[ "$count" -lt 2 ]]; then
-    echo "not json"
-  else
-    echo '{"server":{"status":"running","ip":"172.16.0.1"}}'
-  fi
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/servers/1" "running" \\
-  "d['server']['status']" "d['server']['ip']" \\
-  SERVER_IP "Server" 5
-
-echo "IP=$SERVER_IP"
-rm -f "$COUNTER_FILE"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("IP=172.16.0.1");
-    });
-  });
-
-  describe("logging output", () => {
-    it("should log status progress during polling", () => {
-      const result = runBash(`
-COUNTER_FILE=$(mktemp)
-echo "0" > "$COUNTER_FILE"
-
-mock_api() {
-  local count
-  count=$(cat "$COUNTER_FILE")
-  count=$((count + 1))
-  echo "$count" > "$COUNTER_FILE"
-
-  if [[ "$count" -lt 2 ]]; then
-    echo '{"vm":{"state":"booting","ip":""}}'
-  else
-    echo '{"vm":{"state":"ready","ip":"10.0.0.9"}}'
-  fi
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/vms/1" "ready" \\
-  "d['vm']['state']" "d['vm']['ip']" \\
-  VM_IP "VM" 5
-
-rm -f "$COUNTER_FILE"
-`);
-      expect(result.exitCode).toBe(0);
-      // Check for status logging in stderr
-      expect(result.stderr).toContain("Waiting for VM to become ready");
-      expect(result.stderr).toContain("VM status: booting");
-    });
-
-    it("should log success with IP address", () => {
-      const result = runBash(`
-mock_api() {
-  echo '{"instance":{"status":"active","ip":"10.0.0.1"}}'
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/instances/1" "active" \\
-  "d['instance']['status']" "d['instance']['ip']" \\
-  IP "Instance" 1
-
-echo "IP=$IP"
-`);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("IP=10.0.0.1");
-    });
-
-    it("should log helpful error message on timeout", () => {
-      const result = runBash(`
-mock_api() {
-  echo '{"instance":{"status":"pending","ip":""}}'
-}
-
-INSTANCE_STATUS_POLL_DELAY=0
-generic_wait_for_instance mock_api "/instances/1" "active" \\
-  "d['instance']['status']" "d['instance']['ip']" \\
-  IP "MyInstance" 2
-`);
-      expect(result.exitCode).not.toBe(0);
-      expect(result.stderr).toContain("MyInstance did not become active");
-      expect(result.stderr).toContain("retry");
-    });
-  });
-});
+// _extract_json_field tests are in shared-common-json-extraction.test.ts
+// generic_wait_for_instance tests are in shared-common-ssh-helpers.test.ts
