+++ cli/src/__tests__/shared-common-input-validation.test.ts
@@ -632,181 +632,6 @@ describe("_display_and_select", () => {
 // by verifying what validated_read would accept/reject through
 // the validators themselves.
 
-describe("validated_read validator contract", () => {
-  describe("validate_api_token accepts valid tokens", () => {
-    it("should accept standard API key format", () => {
-      const result = runBash('validate_api_token "sk-or-v1-abc123def456" && echo OK');
-      expect(result.stdout).toBe("OK");
-    });
-
-    it("should accept token with underscores", () => {
-      const result = runBash('validate_api_token "my_api_token_123" && echo OK');
-      expect(result.stdout).toBe("OK");
-    });
-
-    it("should accept token with dots", () => {
-      const result = runBash('validate_api_token "token.with.dots" && echo OK');
-      expect(result.stdout).toBe("OK");
-    });
-  });
-
-  describe("validate_api_token rejects dangerous tokens", () => {
-    it("should reject token with semicolons", () => {
-      const result = runBash('validate_api_token "token;injection"');
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject token with single quotes", () => {
-      const result = runBash("validate_api_token \"token'inject\"");
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject token with double quotes", () => {
-      const result = runBash('validate_api_token "token\\"inject"');
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject token with dollar sign", () => {
-      const result = runBash('validate_api_token "token\\$inject"');
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject token with pipe", () => {
-      const result = runBash('validate_api_token "token|cmd"');
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject token with ampersand", () => {
-      const result = runBash('validate_api_token "token&bg"');
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject empty token", () => {
-      const result = runBash('validate_api_token ""');
-      expect(result.exitCode).not.toBe(0);
-    });
-  });
-
-  describe("validate_region_name via validated_read contract", () => {
-    it("should accept standard AWS-style region", () => {
-      const result = runBash('validate_region_name "us-east-1" && echo OK');
-      expect(result.stdout).toBe("OK");
-    });
-
-    it("should accept region with underscores", () => {
-      const result = runBash('validate_region_name "eu_west_1" && echo OK');
-      expect(result.stdout).toBe("OK");
-    });
-
-    it("should accept single-word region", () => {
-      const result = runBash('validate_region_name "london" && echo OK');
-      expect(result.stdout).toBe("OK");
-    });
-
-    it("should reject region with spaces", () => {
-      const result = runBash('validate_region_name "us east 1"');
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject region longer than 63 chars", () => {
-      const longRegion = "a".repeat(64);
-      const result = runBash(`validate_region_name "${longRegion}"`);
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject empty region", () => {
-      const result = runBash('validate_region_name ""');
-      expect(result.exitCode).not.toBe(0);
-    });
-  });
-
-  describe("validate_resource_name via validated_read contract", () => {
-    it("should accept resource name with dots", () => {
-      const result = runBash('validate_resource_name "e2.micro" && echo OK');
-      expect(result.stdout).toBe("OK");
-    });
-
-    it("should accept resource name with underscores and hyphens", () => {
-      const result = runBash('validate_resource_name "cx11_ssd-fast" && echo OK');
-      expect(result.stdout).toBe("OK");
-    });
-
-    it("should reject resource name with semicolons", () => {
-      const result = runBash('validate_resource_name "type;injection"');
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject resource name with spaces", () => {
-      const result = runBash('validate_resource_name "big server"');
-      expect(result.exitCode).not.toBe(0);
-    });
-
-    it("should reject empty resource name", () => {
-      const result = runBash('validate_resource_name ""');
-      expect(result.exitCode).not.toBe(0);
-    });
-  });
-});
-
-// ── Integration: get_validated_server_name boundary tests ──────────────────
-
-describe("get_validated_server_name boundaries", () => {
-  it("should accept exactly 3 characters (boundary)", () => {
-    const result = runBash(
-      'get_validated_server_name "N" "Enter: "',
-      { env: { N: "abc" } }
-    );
-    expect(result.exitCode).toBe(0);
-    expect(result.stdout).toBe("abc");
-  });
-
-  it("should reject exactly 2 characters (below boundary)", () => {
-    const result = runBash(
-      'get_validated_server_name "N" "Enter: "',
-      { env: { N: "ab" } }
-    );
-    expect(result.exitCode).not.toBe(0);
-  });
-
-  it("should accept exactly 63 characters (boundary)", () => {
-    const name = "a".repeat(63);
-    const result = runBash(
-      'get_validated_server_name "N" "Enter: "',
-      { env: { N: name } }
-    );
-    expect(result.exitCode).toBe(0);
-  });
-
-  it("should reject exactly 64 characters (above boundary)", () => {
-    const name = "a".repeat(64);
-    const result = runBash(
-      'get_validated_server_name "N" "Enter: "',
-      { env: { N: name } }
-    );
-    expect(result.exitCode).not.toBe(0);
-  });
-
-  it("should accept name with dash in middle", () => {
-    const result = runBash(
-      'get_validated_server_name "N" "Enter: "',
-      { env: { N: "a-b" } }
-    );
-    expect(result.exitCode).toBe(0);
-  });
-
-  it("should reject name that is only dashes", () => {
-    const result = runBash(
-      'get_validated_server_name "N" "Enter: "',
-      { env: { N: "---" } }
-    );
-    expect(result.exitCode).not.toBe(0);
-  });
-
-  it("should reject name starting and ending with dashes", () => {
-    const result = runBash(
-      'get_validated_server_name "N" "Enter: "',
-      { env: { N: "-abc-" } }
-    );
-    expect(result.exitCode).not.toBe(0);
-  });
-});
+// Standalone validator tests (validate_api_token, validate_region_name, validate_resource_name)
+// are in shared-common-validators.test.ts
+// get_validated_server_name boundary tests are covered above
