+++ cli/src/__tests__/shared-common-json-extraction.test.ts
@@ -274,308 +274,4 @@ describe("_extract_json_field", () => {
   });
 });
 
-// ── extract_api_error_message ─────────────────────────────────────────────
-
-describe("extract_api_error_message", () => {
-  describe("standard error field: message", () => {
-    it("should extract top-level 'message' field", () => {
-      const result = runBash(`
-        extract_api_error_message '{"message": "Rate limit exceeded"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Rate limit exceeded");
-    });
-
-    it("should prefer error.message over top-level message", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": {"message": "inner error"}, "message": "outer error"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("inner error");
-    });
-  });
-
-  describe("standard error field: error (string)", () => {
-    it("should extract 'error' when it is a string", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": "unauthorized"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("unauthorized");
-    });
-  });
-
-  describe("standard error field: error.message", () => {
-    it("should extract error.message from nested error object", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": {"message": "Invalid API key"}}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Invalid API key");
-    });
-  });
-
-  describe("standard error field: error.error_message", () => {
-    it("should extract error.error_message", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": {"error_message": "Quota exceeded"}}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Quota exceeded");
-    });
-  });
-
-  describe("standard error field: reason", () => {
-    it("should extract 'reason' field", () => {
-      const result = runBash(`
-        extract_api_error_message '{"reason": "Server capacity full"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Server capacity full");
-    });
-  });
-
-  describe("fallback behavior", () => {
-    it("should return default fallback for invalid JSON", () => {
-      const result = runBash(`
-        extract_api_error_message 'not valid json'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Unknown error");
-    });
-
-    it("should return custom fallback for invalid JSON", () => {
-      const result = runBash(`
-        extract_api_error_message 'not valid json' 'Custom fallback'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Custom fallback");
-    });
-
-    it("should return fallback when JSON has no recognized error fields", () => {
-      const result = runBash(`
-        extract_api_error_message '{"status": 500, "data": null}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Unknown error");
-    });
-
-    it("should return custom fallback when no recognized fields", () => {
-      const result = runBash(`
-        extract_api_error_message '{"status": 500}' 'API returned 500'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("API returned 500");
-    });
-
-    it("should return fallback for empty JSON object", () => {
-      const result = runBash(`
-        extract_api_error_message '{}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Unknown error");
-    });
-
-    it("should return fallback for empty string input", () => {
-      const result = runBash(`
-        extract_api_error_message ''
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Unknown error");
-    });
-  });
-
-  describe("field priority order", () => {
-    it("should prefer error.message over error string", () => {
-      // When error is a dict with message, that takes priority
-      const result = runBash(`
-        extract_api_error_message '{"error": {"message": "detailed"}, "reason": "generic"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("detailed");
-    });
-
-    it("should prefer error.error_message when error.message is absent", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": {"error_message": "specific"}, "message": "generic"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("specific");
-    });
-
-    it("should fall back to message when error is not a dict", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": 42, "message": "top-level message"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("top-level message");
-    });
-
-    it("should fall back to reason when message is absent", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": 42, "reason": "the reason"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("the reason");
-    });
-
-    it("should use error string when no dict or other fields", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": "simple string error"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("simple string error");
-    });
-  });
-
-  describe("real-world cloud provider error responses", () => {
-    it("should parse Hetzner error format", () => {
-      const json = '{"error": {"message": "server limit exceeded", "code": "limit_exceeded"}}';
-      const result = runBash(`
-        extract_api_error_message '${json}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("server limit exceeded");
-    });
-
-    it("should parse DigitalOcean error format", () => {
-      const json = '{"id": "unauthorized", "message": "Unable to authenticate you"}';
-      const result = runBash(`
-        extract_api_error_message '${json}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Unable to authenticate you");
-    });
-
-    it("should parse Vultr error format", () => {
-      const json = '{"error": "Invalid API token", "status": 401}';
-      const result = runBash(`
-        extract_api_error_message '${json}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Invalid API token");
-    });
-
-    it("should parse Contabo error format", () => {
-      const json = '{"error": {"message": "Insufficient credits"}, "statusCode": 402}';
-      const result = runBash(`
-        extract_api_error_message '${json}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Insufficient credits");
-    });
-
-    it("should parse Linode error format", () => {
-      const json = '{"errors": [{"reason": "Not found"}]}';
-      const result = runBash(`
-        extract_api_error_message '${json}' 'Linode API error'
-      `);
-      // Linode uses errors[] array, not a field extract_api_error_message checks.
-      // Should fall back to custom fallback.
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Linode API error");
-    });
-
-    it("should handle HTML error page gracefully", () => {
-      const result = runBash(`
-        extract_api_error_message '<html><body>502 Bad Gateway</body></html>' 'Server error'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Server error");
-    });
-
-    it("should handle rate limit response", () => {
-      const json = '{"message": "API rate limit exceeded. Try again in 30 seconds."}';
-      const result = runBash(`
-        extract_api_error_message '${json}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("API rate limit exceeded. Try again in 30 seconds.");
-    });
-  });
-
-  describe("edge cases", () => {
-    it("should handle error field set to null", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": null, "message": "null error"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("null error");
-    });
-
-    it("should handle error field set to empty string", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": "", "message": "empty error field"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("empty error field");
-    });
-
-    it("should handle error field set to empty dict", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": {}, "message": "dict but empty"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("dict but empty");
-    });
-
-    it("should handle error dict with message set to empty string", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": {"message": ""}, "reason": "fallback reason"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("fallback reason");
-    });
-
-    it("should handle message field with special characters", () => {
-      const result = runBash(`
-        extract_api_error_message '{"message": "Error: can'\\''t connect to host (port 443)"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toContain("connect to host");
-    });
-
-    it("should handle very long error messages", () => {
-      const longMsg = "x".repeat(500);
-      const result = runBash(`
-        extract_api_error_message '{"message": "${longMsg}"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout.length).toBe(500);
-    });
-
-    it("should handle JSON with only numeric fields", () => {
-      const result = runBash(`
-        extract_api_error_message '{"code": 403, "status": 403}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("Unknown error");
-    });
-
-    it("should handle JSON array input", () => {
-      const result = runBash(`
-        extract_api_error_message '[{"error": "bad"}]' 'array input'
-      `);
-      expect(result.exitCode).toBe(0);
-      // JSON array root doesn't have .get(), should fall back
-      expect(result.stdout).toBe("array input");
-    });
-
-    it("should handle error with both message and error_message", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": {"message": "primary", "error_message": "secondary"}}'
-      `);
-      expect(result.exitCode).toBe(0);
-      // message should take priority over error_message
-      expect(result.stdout).toBe("primary");
-    });
-
-    it("should handle boolean error field", () => {
-      const result = runBash(`
-        extract_api_error_message '{"error": true, "message": "bool error"}'
-      `);
-      expect(result.exitCode).toBe(0);
-      expect(result.stdout).toBe("bool error");
-    });
-  });
-});
+// extract_api_error_message tests are in shared-common-error-polling.test.ts
