+++ cli/src/__tests__/shared-common-logging-utils.test.ts
@@ -471,81 +471,7 @@ describe("calculate_retry_backoff", () => {
   });
 });
 
-// ── generate_ssh_key_if_missing ─────────────────────────────────────────────
-
-describe("generate_ssh_key_if_missing", () => {
-  it("should create key when it does not exist", () => {
-    const keyPath = join(testDir, "test_key");
-    const result = runBash(`generate_ssh_key_if_missing "${keyPath}"`);
-    expect(result.exitCode).toBe(0);
-    expect(existsSync(keyPath)).toBe(true);
-    expect(existsSync(`${keyPath}.pub`)).toBe(true);
-  });
-
-  it("should not overwrite existing key", () => {
-    const keyPath = join(testDir, "existing_key");
-    writeFileSync(keyPath, "existing-key-content");
-
-    const result = runBash(`generate_ssh_key_if_missing "${keyPath}"`);
-    expect(result.exitCode).toBe(0);
-    expect(readFileSync(keyPath, "utf-8")).toBe("existing-key-content");
-  });
-
-  it("should create parent directories if needed", () => {
-    const keyPath = join(testDir, "nested", "dir", "key");
-    const result = runBash(`generate_ssh_key_if_missing "${keyPath}"`);
-    expect(result.exitCode).toBe(0);
-    expect(existsSync(keyPath)).toBe(true);
-  });
-
-  it("should generate ed25519 key type", () => {
-    const keyPath = join(testDir, "ed_key");
-    runBash(`generate_ssh_key_if_missing "${keyPath}"`);
-    const pubContent = readFileSync(`${keyPath}.pub`, "utf-8");
-    expect(pubContent).toContain("ssh-ed25519");
-  });
-
-  it("should generate key with no passphrase", () => {
-    const keyPath = join(testDir, "nopw_key");
-    // Key should be usable without passphrase
-    const result = runBash(`
-      generate_ssh_key_if_missing "${keyPath}"
-      ssh-keygen -y -f "${keyPath}" > /dev/null 2>&1
-    `);
-    expect(result.exitCode).toBe(0);
-  });
-});
-
-// ── get_ssh_fingerprint ─────────────────────────────────────────────────────
-
-describe("get_ssh_fingerprint", () => {
-  it("should return MD5 fingerprint of a public key", () => {
-    const keyPath = join(testDir, "fp_key");
-    runBash(`ssh-keygen -t ed25519 -f "${keyPath}" -N "" -q`);
-
-    const result = runBash(`get_ssh_fingerprint "${keyPath}.pub"`);
-    expect(result.exitCode).toBe(0);
-    // MD5 fingerprint format: xx:xx:xx:xx:...
-    expect(result.stdout).toMatch(/^[0-9a-f]{2}(:[0-9a-f]{2}){15}$/);
-  });
-
-  it("should not include MD5: prefix in output", () => {
-    const keyPath = join(testDir, "fp_key2");
-    runBash(`ssh-keygen -t ed25519 -f "${keyPath}" -N "" -q`);
-
-    const result = runBash(`get_ssh_fingerprint "${keyPath}.pub"`);
-    expect(result.stdout).not.toContain("MD5:");
-  });
-
-  it("should return consistent fingerprint for same key", () => {
-    const keyPath = join(testDir, "fp_key3");
-    runBash(`ssh-keygen -t ed25519 -f "${keyPath}" -N "" -q`);
-
-    const result1 = runBash(`get_ssh_fingerprint "${keyPath}.pub"`);
-    const result2 = runBash(`get_ssh_fingerprint "${keyPath}.pub"`);
-    expect(result1.stdout).toBe(result2.stdout);
-  });
-});
+// generate_ssh_key_if_missing and get_ssh_fingerprint tests are in shared-common-ssh-key-lifecycle.test.ts
 
 // ── logging functions ───────────────────────────────────────────────────────
