+++ cli/src/__tests__/shared-common-ssh-helpers.test.ts
@@ -613,138 +613,4 @@ generic_wait_for_instance mock_api "/s/1" "running" \\
   });
 });
 
-// ── extract_api_error_message ────────────────────────────────────────────────
-
-describe("extract_api_error_message", () => {
-  it("should extract message from {message: ...} pattern", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message '{"message": "Server not found"}'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("Server not found");
-  });
-
-  it("should extract from {error: {message: ...}} nested pattern", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message '{"error": {"message": "Rate limit exceeded"}}'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("Rate limit exceeded");
-  });
-
-  it("should extract from {error: {error_message: ...}} pattern", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message '{"error": {"error_message": "Invalid token"}}'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("Invalid token");
-  });
-
-  it("should extract from {reason: ...} pattern", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message '{"reason": "Unauthorized"}'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("Unauthorized");
-  });
-
-  it("should extract from {error: 'string'} pattern", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message '{"error": "Bad request"}'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("Bad request");
-  });
-
-  it("should use fallback when JSON is invalid", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message 'not json at all' 'Custom fallback'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("Custom fallback");
-  });
-
-  it("should use default fallback 'Unknown error' when no fallback specified", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message 'broken json'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("Unknown error");
-  });
-
-  it("should use fallback when JSON has no recognized error fields", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message '{"data": "success", "status": 200}' 'No error found'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("No error found");
-  });
-
-  it("should prefer error.message over top-level message", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message '{"error": {"message": "nested"}, "message": "top-level"}'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("nested");
-  });
-
-  it("should handle empty JSON object with fallback", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message '{}' 'Empty response'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("Empty response");
-  });
-
-  it("should handle error message with special characters", () => {
-    const { stdout, exitCode } = runBash(
-      `extract_api_error_message '{"message": "Error: can'\\''t connect to host (port 443)"}'`
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toContain("can't connect");
-  });
-});
-
-// ── ensure_ssh_key_with_provider (argument construction) ─────────────────────
-
-describe("ensure_ssh_key_with_provider function exists", () => {
-  it("should be defined in shared/common.sh", () => {
-    const { exitCode, stdout } = runBash(
-      'type -t ensure_ssh_key_with_provider'
-    );
-    expect(exitCode).toBe(0);
-    expect(stdout).toBe("function");
-  });
-});
-
-// ── SSH_USER default behavior ────────────────────────────────────────────────
-
-describe("SSH_USER default behavior across helpers", () => {
-  it("should use root as default for all SSH helpers when SSH_USER is unset", () => {
-    createMockCommand("ssh", 'echo "$@"');
-    createMockCommand("scp", 'echo "$@"');
-
-    const run = runBash('SSH_OPTS=""\nssh_run_server "10.0.0.1" "cmd"', { useMockPath: true });
-    expect(run.stdout).toContain("root@10.0.0.1");
-
-    const upload = runBash('SSH_OPTS=""\nssh_upload_file "10.0.0.1" "/a" "/b"', { useMockPath: true });
-    expect(upload.stdout).toContain("root@10.0.0.1");
-
-    const interactive = runBash('SSH_OPTS=""\nssh_interactive_session "10.0.0.1" "sh"', { useMockPath: true });
-    expect(interactive.stdout).toContain("root@10.0.0.1");
-  });
-
-  it("should use custom SSH_USER consistently across all helpers", () => {
-    createMockCommand("ssh", 'echo "$@"');
-    createMockCommand("scp", 'echo "$@"');
-
-    const run = runBash('SSH_OPTS=""\nSSH_USER=custom\nssh_run_server "10.0.0.1" "cmd"', { useMockPath: true });
-    expect(run.stdout).toContain("custom@10.0.0.1");
-
-    const upload = runBash('SSH_OPTS=""\nSSH_USER=custom\nssh_upload_file "10.0.0.1" "/a" "/b"', { useMockPath: true });
-    expect(upload.stdout).toContain("custom@10.0.0.1");
-
-    const interactive = runBash('SSH_OPTS=""\nSSH_USER=custom\nssh_interactive_session "10.0.0.1" "sh"', { useMockPath: true });
-    expect(interactive.stdout).toContain("custom@10.0.0.1");
-  });
-});
+// extract_api_error_message tests are in shared-common-error-polling.test.ts
