+++ cli/src/__tests__/shared-common-ssh-key-lifecycle.test.ts
@@ -620,146 +620,5 @@ describe("ensure_ssh_key_with_provider", () => {
   });
 });
 
-// ── extract_ssh_key_ids ──────────────────────────────────────────────────
-
-describe("extract_ssh_key_ids edge cases", () => {
-  it("should extract IDs from standard ssh_keys response", () => {
-    const json = JSON.stringify({
-      ssh_keys: [{ id: 123 }, { id: 456 }, { id: 789 }],
-    });
-    const result = runBash(
-      `echo '${json}' | python3 -c "
-import json, sys
-data = json.loads(sys.stdin.read())
-ids = [k['id'] for k in data.get('ssh_keys', [])]
-print(json.dumps(ids))
-"`
-    );
-    expect(result.exitCode).toBe(0);
-    expect(JSON.parse(result.stdout)).toEqual([123, 456, 789]);
-  });
-
-  it("should return empty array when key_field is empty", () => {
-    const json = JSON.stringify({ ssh_keys: [] });
-    const result = runBash(`extract_ssh_key_ids '${json}' "ssh_keys"`);
-    expect(result.exitCode).toBe(0);
-    expect(JSON.parse(result.stdout)).toEqual([]);
-  });
-
-  it("should work with custom key field name (data)", () => {
-    const json = JSON.stringify({
-      data: [{ id: "key-1" }, { id: "key-2" }],
-    });
-    const result = runBash(`extract_ssh_key_ids '${json}' "data"`);
-    expect(result.exitCode).toBe(0);
-    expect(JSON.parse(result.stdout)).toEqual(["key-1", "key-2"]);
-  });
-
-  it("should return empty array when key_field is missing from response", () => {
-    const json = JSON.stringify({ other_field: [{ id: 1 }] });
-    const result = runBash(`extract_ssh_key_ids '${json}' "ssh_keys"`);
-    expect(result.exitCode).toBe(0);
-    expect(JSON.parse(result.stdout)).toEqual([]);
-  });
-
-  it("should handle numeric and string IDs", () => {
-    const json = JSON.stringify({
-      ssh_keys: [{ id: 42 }, { id: "abc-def" }],
-    });
-    const result = runBash(`extract_ssh_key_ids '${json}' "ssh_keys"`);
-    expect(result.exitCode).toBe(0);
-    const ids = JSON.parse(result.stdout);
-    expect(ids).toContain(42);
-    expect(ids).toContain("abc-def");
-  });
-
-  it("should default to ssh_keys field when no field specified", () => {
-    const json = JSON.stringify({
-      ssh_keys: [{ id: 100 }],
-    });
-    const result = runBash(`extract_ssh_key_ids '${json}'`);
-    expect(result.exitCode).toBe(0);
-    expect(JSON.parse(result.stdout)).toEqual([100]);
-  });
-
-  it("should handle single key in response", () => {
-    const json = JSON.stringify({
-      ssh_keys: [{ id: 1 }],
-    });
-    const result = runBash(`extract_ssh_key_ids '${json}' "ssh_keys"`);
-    expect(result.exitCode).toBe(0);
-    expect(JSON.parse(result.stdout)).toEqual([1]);
-  });
-});
-
-// ── check_ssh_key_by_fingerprint ─────────────────────────────────────────
-
-describe("check_ssh_key_by_fingerprint edge cases", () => {
-  it("should succeed when fingerprint is found in API response", () => {
-    const result = runBash(`
-      mock_api() {
-        echo '{"ssh_keys": [{"fingerprint": "aa:bb:cc:dd:ee:ff"}]}'
-      }
-      check_ssh_key_by_fingerprint mock_api "/ssh_keys" "aa:bb:cc:dd:ee:ff"
-    `);
-    expect(result.exitCode).toBe(0);
-  });
-
-  it("should fail when fingerprint is not in API response", () => {
-    const result = runBash(`
-      mock_api() {
-        echo '{"ssh_keys": [{"fingerprint": "11:22:33:44:55:66"}]}'
-      }
-      check_ssh_key_by_fingerprint mock_api "/ssh_keys" "aa:bb:cc:dd:ee:ff"
-    `);
-    expect(result.exitCode).not.toBe(0);
-  });
-
-  it("should match partial fingerprint strings", () => {
-    const result = runBash(`
-      mock_api() {
-        echo '{"ssh_keys": [{"fingerprint": "aa:bb:cc:dd:ee:ff:11:22:33:44:55:66:77:88:99:00"}]}'
-      }
-      check_ssh_key_by_fingerprint mock_api "/ssh_keys" "aa:bb:cc:dd"
-    `);
-    expect(result.exitCode).toBe(0);
-  });
-
-  it("should fail when API returns empty key list", () => {
-    const result = runBash(`
-      mock_api() {
-        echo '{"ssh_keys": []}'
-      }
-      check_ssh_key_by_fingerprint mock_api "/ssh_keys" "aa:bb:cc"
-    `);
-    expect(result.exitCode).not.toBe(0);
-  });
-
-  it("should pass correct endpoint to API function", () => {
-    const result = runBash(`
-      mock_api() {
-        if [[ "\$2" == "/v2/account/keys" ]]; then
-          echo '{"ssh_keys": [{"fingerprint": "test-fp"}]}'
-        else
-          echo '{"ssh_keys": []}'
-        fi
-      }
-      check_ssh_key_by_fingerprint mock_api "/v2/account/keys" "test-fp"
-    `);
-    expect(result.exitCode).toBe(0);
-  });
-
-  it("should use GET method when calling API function", () => {
-    const result = runBash(`
-      mock_api() {
-        if [[ "\$1" == "GET" ]]; then
-          echo '{"ssh_keys": [{"fingerprint": "my-fp"}]}'
-        else
-          echo '{"ssh_keys": []}'
-        fi
-      }
-      check_ssh_key_by_fingerprint mock_api "/ssh_keys" "my-fp"
-    `);
-    expect(result.exitCode).toBe(0);
-  });
-});
+// extract_ssh_key_ids tests are in shared-common-untested-helpers.test.ts
+// check_ssh_key_by_fingerprint tests are in shared-common-env-inject.test.ts
